Refactor NodeLogHandler into view_logs.py
